Fix override_footnotestyle() issues
On a fresh install, without setting colors for
wp-bigfoot, the option names `wpbf-bgcolor` and
`wpbf-fgcolor` are unknown and generate PHP
“undefined index” notices.

If only one color is set the other CSS statement
has no value and this results in an CSS error.
This fix only shows a CSS statement if a color is
set.

Move the STYLE tag into the HEAD of the HTML via
`wp_head` hook.